+++ build_runner_core/lib/src/asset/finalized_reader.dart
@@ -14,22 +14,24 @@ import 'package:glob/glob.dart';
 import '../asset_graph/graph.dart';
 import '../asset_graph/node.dart';
 import '../asset_graph/optional_output_tracker.dart';
+import '../package_graph/target_graph.dart';
 
 /// An [AssetReader] which ignores deleted files.
 class FinalizedReader implements AssetReader {
   final AssetReader _delegate;
   final AssetGraph _assetGraph;
+  final TargetGraph _targetGraph;
   OptionalOutputTracker? _optionalOutputTracker;
   final String _rootPackage;
   final List<BuildPhase> _buildPhases;
 
   void reset(Set<String> buildDirs, Set<BuildFilter> buildFilters) {
     _optionalOutputTracker = OptionalOutputTracker(
-        _assetGraph, buildDirs, buildFilters, _buildPhases);
+        _assetGraph, _targetGraph, buildDirs, buildFilters, _buildPhases);
   }
 
-  FinalizedReader(
-      this._delegate, this._assetGraph, this._buildPhases, this._rootPackage);
+  FinalizedReader(this._delegate, this._assetGraph, this._targetGraph,
+      this._buildPhases, this._rootPackage);
 
   /// Returns a reason why [id] is not readable, or null if it is readable.
   Future<UnreadableReason?> unreadableReason(AssetId id) async {