+++ build_runner_core/lib/src/asset/lru_cache.dart
@@ -27,6 +27,7 @@ class LruCache<K, V> {
   }
 
   void operator []=(K key, V value) {
+    remove(key);
     var entry = _Link(key, value, _computeWeight(value));
     // Don't cache at all if above the individual weight max.
     if (entry.weight > _individualWeightMax) {