+++ build_vm_compilers/CHANGELOG.md
@@ -1,4 +1,12 @@
-## 1.0.12-dev
+## 1.0.13
+
+- Migrate off deprecated analyzer apis.
+- Update min sdk constraint to 2.18.0.
+- Expand the pubspec description.
+
+## 1.0.12
+
+- Allow the latest `package:analyzer`.
 
 ## 1.0.11