+++ build_web_compilers/CHANGELOG.md
@@ -1,4 +1,12 @@
-## 3.2.6-dev
+## 3.2.7
+
+- Migrate off deprecated analyzer apis.
+- Update min sdk constraint to 2.18.0.
+- Expand the pubspec description.
+
+## 3.2.6
+
+- Allow the latest `package:analyzer`.
 
 ## 3.2.5